+++ src/main/java/net/kaikk/mc/bcl/config/Config.java
@@ -75,7 +75,7 @@ public class Config implements Configurable {
     public void populate() {
         get().getNode("ServerName").setValue("aServer").setComment("Unique name of server");
         get().getNode("MaxHoursOffline").setValue(72).setComment("Time in hours before player's chunkloaders become inactive.");
-        get().getNode("DataStore").setValue("MySQL").setComment("Currently the only DataStore...");
+        get().getNode("DataStore").setValue("MySQL").setComment("The only DataStore available is MySQL atm!");
         get().getNode("DefaultChunksAmount", "World").setValue(0);
         get().getNode("DefaultChunksAmount", "Personal").setValue(0);
         get().getNode("MaxChunksAmount", "World").setValue(250);